test: cobertura dos endpoints de todos os controllers e tratamento global de exceptions

+++ Backend/src/main/java/br/com/juridico/totvs/fullstack/Backend/services/impls/PaisServiceImpl.java
@@ -1,6 +1,7 @@
 package br.com.juridico.totvs.fullstack.Backend.services.impls;
 
 import br.com.juridico.totvs.fullstack.Backend.domains.Pais;
+import br.com.juridico.totvs.fullstack.Backend.excptions.RecursoNaoEncontradoException;
 import br.com.juridico.totvs.fullstack.Backend.repositories.PaisRepository;
 import br.com.juridico.totvs.fullstack.Backend.services.dtos.pais.PaisCreateUpdateDTO;
 import br.com.juridico.totvs.fullstack.Backend.services.dtos.pais.PaisDTO;
@@ -36,7 +37,11 @@ public class PaisServiceImpl implements PaisService {
     @Override
     public PaisDTO update(Long id, PaisCreateUpdateDTO dto) {
         Pais pais = paisRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("País não encontrado"));
+                .orElseThrow(() -> new RecursoNaoEncontradoException(
+                        "País com ID "
+                                + id
+                                + " não encontrado")
+                );
 
         pais.setNome(dto.getNome());
         pais.setSigla(dto.getSigla());
@@ -49,13 +54,22 @@ public class PaisServiceImpl implements PaisService {
 
     @Override
     public void delete(Long id) {
+        if (!paisRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException(
+                    "País com ID "
+                            + id
+                            + " não encontrado");
+        }
         paisRepository.deleteById(id);
     }
 
     @Override
     public PaisDTO getPaisById(Long id) {
         Pais pais = paisRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("País não encontrado"));
+                .orElseThrow(() -> new RecursoNaoEncontradoException(
+                        "País com ID "
+                                + id
+                                + " não encontrado"));
         return new PaisDTO(pais);
 
     }
@@ -76,7 +90,9 @@ public class PaisServiceImpl implements PaisService {
     @Override
     public List<PontoTuristicoDTO> getPontosTuristicos(Long id) {
         Pais pais = paisRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("País não encontrado"));
+                .orElseThrow(() -> new RecursoNaoEncontradoException("País com ID "
+                        + id
+                        + " não encontrado"));
 
         return pais.getPontosTuristicos().stream()
                 .map(PontoTuristicoDTO::new)
